fix(ui): project roles has not been updated
