+++ ui/src/features/projects/members/members.mutation.ts
@@ -6,6 +6,8 @@ import {
 } from '@matrixhub/api-ts/v1alpha1/project.pb'
 import { mutationOptions } from '@tanstack/react-query'
 
+import { authKeys } from '@/features/auth/auth.query'
+
 import { memberKeys } from './members.query'
 
 export function addMemberMutationOptions() {
@@ -13,7 +15,7 @@ export function addMemberMutationOptions() {
     mutationFn: (input: AddProjectMemberWithRoleRequest) =>
       Projects.AddProjectMemberWithRole(input),
     meta: {
-      invalidates: [memberKeys.lists()],
+      invalidates: [memberKeys.lists(), authKeys.projectRoles()],
     },
   })
 }
@@ -23,7 +25,7 @@ export function updateMemberRoleMutationOptions() {
     mutationFn: (input: UpdateProjectMemberRoleRequest) =>
       Projects.UpdateProjectMemberRole(input),
     meta: {
-      invalidates: [memberKeys.lists()],
+      invalidates: [memberKeys.lists(), authKeys.projectRoles()],
     },
   })
 }
@@ -33,7 +35,7 @@ export function removeMembersMutationOptions() {
     mutationFn: (input: RemoveProjectMembersRequest) =>
       Projects.RemoveProjectMembers(input),
     meta: {
-      invalidates: [memberKeys.lists()],
+      invalidates: [memberKeys.lists(), authKeys.projectRoles()],
     },
   })
 }
