+++ ui/src/routes/(auth)/(app)/projects/$projectId/route.tsx
@@ -13,7 +13,7 @@ import {
 } from '@tanstack/react-router'
 import { useTranslation } from 'react-i18next'
 
-import { useProjectRole } from '@/context/project-role-context'
+import { useProjectRole } from '@/features/auth/useProjectRole'
 import { projectDetailQueryOptions } from '@/features/projects/projects.query'
 
 import { Route as ProjectMembersRoute } from './members'
