+++ ui/src/routes/(auth)/(app)/projects_.$projectId/datasets.$datasetId/route.tsx
@@ -19,8 +19,7 @@ import {
 } from '@tanstack/react-router'
 import { useTranslation } from 'react-i18next'
 
-import { projectRolesQueryOptions } from '@/features/auth/auth.query'
-import { queryClient } from '@/queryClient'
+import { useProjectRole } from '@/features/auth/useProjectRole'
 import { ResourceDetailHeader } from '@/shared/components/ResourceDetailHeader'
 
 import { Route as DatasetSettingsRoute } from './settings'
@@ -46,13 +45,11 @@ export const Route = createFileRoute(
 )({
   component: DatasetLayout,
   loader: async () => {
-    const projectRoles = await queryClient.ensureQueryData(projectRolesQueryOptions())
     // TODO fetch dataset detail data, if the project is private and user has no access, will throw error
     // throw new NotFoundRouteError()
 
     return {
       dataset: MOCK_DATA,
-      projectRoles: projectRoles,
     }
   },
 })
@@ -64,11 +61,8 @@ function DatasetLayout() {
     projectId, datasetId,
   } = Route.useParams()
 
-  const {
-    projectRoles,
-  } = Route.useLoaderData()
-
-  const hasProjectRole = Object.hasOwn(projectRoles.projectRoles ?? {}, projectId)
+  const projectRole = useProjectRole(projectId)
+  const hasProjectRole = projectRole !== undefined
 
   const tabRoutes = linkOptions([
     {
